requirements.txt
fastapi
uvicorn[standard]
python-multipart
boto3
PyPDF2
pikepdf
pypdf
Pillow